Move raw pointer core tests to extern specs
surface now keeps only raw pointer language-level coverage:

- tests/tests/pos/surface/raw_ptr00.rs checks refined *const / *mut
syntax, ptr_size, ptr_offset, and generic T::size_of() in annotations.
- tests/tests/neg/surface/raw_ptr00.rs checks basic raw pointer
refinement subtyping failures.

Core/extern-spec behavior moved to extern_specs:

- raw_ptr01.rs and raw_ptr02.rs moved to flux_core_ptr01.rs /
flux_core_ptr02.rs.
- The old core-dependent raw_ptr00.rs content moved to
flux_core_ptr00.rs.
- This puts std::ptr::read, std::ptr::write, add, sub, and as_mut_ptr
coverage under extern_specs.
