+++ tests/tests/neg/extern_specs/flux_core_ptr02.rs
