+++ tests/tests/pos/extern_specs/flux_core_ptr02.rs
